final commit

+++ app/Http/Controllers/ListingController.php
@@ -65,6 +65,11 @@ class ListingController extends Controller
 
 
     public function update(Request $request,Listing $listing){
+
+        if($listing->user_id!=auth()->id()){
+            abort('403','UNAUTHORIZED ACTION');
+        }
+
         $formfield = $request->validate(
          [
              'title'=>'required',
@@ -92,13 +97,18 @@ class ListingController extends Controller
      }
 
      public function destroy(Listing $listing){
+
+        if($listing->user_id!=auth()->id()){
+            abort('403','UNAUTHORIZED ACTION');
+        }
+
          $listing->delete();
 
          return redirect('/')->with('message','Listing deleted Succesfully');
      }
 
      public function manage(){
-        dd();
+        
         return view(
             'listings.manage',['listings'=>auth()->user()->listings()->get()]
         );
